+++ appTools/ToolDrilling.py
@@ -2750,7 +2750,7 @@ class DrillingUI:
         self.over_z_label.setToolTip(_("The height Z to which the tool will rise in order to avoid\n"
                                        "an interdiction area."))
         self.over_z_entry = FCDoubleSpinner()
-        self.over_z_entry.set_range(0.000, 10000.0000)
+        self.over_z_entry.set_range(-10000.0000, 10000.0000)
         self.over_z_entry.set_precision(self.decimals)
         self.over_z_entry.setObjectName("e_area_overz")